Fix build on non-Linux OSes
